openAuthSession
