+++ packages/client/src/components/communication/index.ts
@@ -1,2 +1,1 @@
-export * from './handleResponse';
 export * from './postRequestToWallet';
