+++ packages/client/src/components/communication/utils/encoding.test.ts
