+++ packages/client/src/components/communication/utils/encoding.ts
