+++ packages/client/src/components/communication/utils/types.ts
